Address initial feedback
